% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{ARdet}
\alias{ARdet}
\title{Determinant for an AR(1) covariance matrix.}
\usage{
ARdet(n, rho, deriv = FALSE)
}
\arguments{
\item{n}{dimensions}

\item{rho}{off-diagonal parameter}

\item{deriv}{logical whether to return the deterimant or the derivative of the determinant}
}
\value{
determinant of an AR(1) covariance matrix
}
\description{
Determinant for an AR(1) covariance matrix.
}
\examples{
\dontrun{
ARdet(toepgenerate(5,.4,deriv=TRUE))
}

}
\keyword{internal}
