% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{CSgenerate}
\alias{CSgenerate}
\title{Generate a compound symmetric correlation matrix}
\usage{
CSgenerate(n, rho)
}
\arguments{
\item{n}{number of dimensions}

\item{rho}{off-diagonal element - a correlation between -1 and 1. Will warn if less than 0.}
}
\value{
returns an \eqn{n x n} matrix with 1 on the diagonal and \code{rho} on the off-diagonal.
}
\description{
Generate a compound symmetric correlation matrix
}
\examples{
CSgenerate(3,.5)
}
