% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm.R
\name{init_matrixmixture}
\alias{init_matrixmixture}
\title{Initializing settings for Matrix Mixture Models}
\usage{
init_matrixmixture(data, prior = NULL, K = length(prior),
  centers = NULL, U = NULL, V = NULL, centermethod = "kmeans",
  varmethod = "identity", model = "normal", init = NULL, ...)
}
\arguments{
\item{data}{array of data}

\item{prior}{prior probability. One of \code{prior} and \code{K} must be
provided. They must be consistent if both provided.}

\item{K}{number of groups}

\item{centers}{either a matrix or an array of matrices for use as the
\code{centers} argument (optional)}

\item{U}{either a matrix or an array of matrices for use as the \code{U}
argument (optional)}

\item{V}{either a matrix or an array of matrices for use as the \code{V}
argument (optional)}

\item{centermethod}{what method to use to generate initial centers.
Current support random start or performing k-means on the vectorized
version for a small number of iterations and then converting back.
By default, if centers are provided, nothing will be done.}

\item{varmethod}{what method to use to choose initial variance matrices.
Currently either identity matrices or the empirical covariance matrix
determined by hard assignment to the nearest centers.
By default, if \code{U} and \code{V} matrices are provided, nothing
will be done.}

\item{model}{whether to use a normal distribution or a t-distribution, not
relevant for more initialization methods.}

\item{init}{(optional) a (possibly partially-formed) list with some of the components
\code{centers}, \code{U}, and \code{V}. The function will complete the
list and fill out missing entries.}

\item{...}{Additional arguments to pass to \code{kmeans()}.}
}
\value{
a list suitable to use as the \code{init} argument in
     \code{matrixmixture}
}
\description{
Providing this will generate a list suitable for use as the \code{init}
argument in the \code{matrixmixture} function. Either provide data
and it will select centers and variance matrices to initialize or
provide initial values and it will format them as expected for the function.
}
