% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{invAR}
\alias{invAR}
\title{Returns the inverse of an AR(1) covariance matrix or its derivative}
\usage{
invAR(n, rho, deriv = FALSE)
}
\arguments{
\item{n}{dimensions of matrix}

\item{rho}{correlation parameter}

\item{deriv}{logical. if TRUE will output the derivative of the inverse matrix.}
}
\value{
Matrix of the inverse of the AR(1) covariance matrix (or its inverse)
}
\description{
Returns the inverse of an AR(1) covariance matrix or its derivative
}
\examples{
\dontrun{
derivinvAR(5,.5, TRUE)
}
}
\keyword{internal}
