% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{invCS}
\alias{invCS}
\title{Inverse of Compound Symmetric Matrix.}
\usage{
invCS(n, rho, deriv = FALSE)
}
\arguments{
\item{n}{dimension of the matrix}

\item{rho}{compound symmetric factor}

\item{deriv}{logical, whether to return the derivative of the inverse.}
}
\value{
inverse matrix or derivative of the inverse matrix.
}
\description{
Inverse of Compound Symmetric Matrix.
}
\examples{
\dontrun{
invCS(3,.25,TRUE)
}
}
\keyword{internal}
