% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\docType{package}
\name{matrixdist}
\alias{matrixdist}
\alias{matrixdist-package}
\title{matrixdist: a package for some random matrix distributions.}
\description{
The matrixdist package current provides for generating
matrix normal random variables, computing matrix normal densities,
and fitting the MLEs for parameters of matrix normal distributions.
In the future, this may expand to matrix t distributions or some
other matrix random variables, at least generation and distribution
if not parameter estimation.
}
