% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm.R
\name{matrixmixture}
\alias{matrixmixture}
\title{Fit a matrix variate mixture model}
\usage{
matrixmixture(x, init = NULL, prior = NULL, K = length(prior),
  iter = 1000, model = "normal", method = NULL, tolerance = 0.1,
  nu = NULL, ..., verbose = 0, miniter = 3)
}
\arguments{
\item{x}{data, \code{p x q x n} array}

\item{init}{a list containing an array of \code{K} of \code{p x q} means,
and optionally \code{p x p} and \code{q x q} positive definite variance
matrices. By default, those are presumed to be identity if not provided.
If \code{init} is missing, it will be provided using the prior or K by
\code{init_matrixmix}.}

\item{prior}{prior for the \code{K} classes, a vector that adds to unity}

\item{K}{number of classes - provide either this or the prior. If this is
provided, the prior will be of equal distribution.}

\item{iter}{maximum number of iterations.}

\item{model}{whether to use the \code{normal} or \code{t} distribution.}

\item{method}{what method to use to fit the distribution. Currently no options.}

\item{tolerance}{convergence criterion, using Aitken acceleration of the
log-likelihood.}

\item{nu}{degrees of freedom parameter}

\item{...}{pass additional arguments to \code{MLmatrixnorm} or \code{MLmatrixt}}

\item{verbose}{whether to print diagnostic output, by default \code{0}. Higher
numbers output more results.}

\item{miniter}{minimum number of iterations}
}
\value{
A list of class \code{MixMatrixModel} containing the following
    components:
\describe{
     \item{\code{prior}}{the prior probabilities used.}
     \item{\code{init}}{the initialization used.}
      \item{\code{K}}{the number of groups} 
      \item{\code{n}}{the number of observations} 
      \item{\code{centers}}{the group means.}
      \item{\code{U}}{the between-row covariance matrices}
      \item{\code{V}}{the between-column covariance matrix}
      \item{\code{posterior}}{the posterior probabilities for each observation}
      \item{\code{pi}}{ the final proportions}
      \item{\code{nu}}{The degrees of freedom parameter if the t distribution
           was used.}
      \item{\code{convergence }}{whether the model converged}
      \item{\code{logLik }}{a vector of the log-likelihoods of each iteration ending in
              the final log-likelihood of the model}
      \item{\code{method }}{the method used}
      \item{\code{call}}{The (matched) function call.}
   }
}
\description{
Fit a matrix variate mixture model
}
\examples{

set.seed(20180221)
A <- rmatrixnorm(30,mean=matrix(0,nrow=3,ncol=4))
B <- rmatrixnorm(30,mean=matrix(2,nrow=3,ncol=4))
C <- array(c(A,B), dim=c(3,4,60))
prior <- c(.5,.5)
init = list(centers = array(c(rep(0,12),rep(2,12)), dim = c(3,4,2)),
             U = array(c(diag(3), diag(3)), dim = c(3,3,2)),
             V = array(c(diag(4), diag(4)), dim = c(4,4,2))
             )
res<-matrixmixture(C, init = init, prior = prior)
print(res) # note: prints head of posterior, not full list
if(interactive()) plot(res)
res<-matrixmixture(C, init = init, prior = prior, model = "t", nu = 5)
if(interactive()) plot(res)
}
