% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{matrixtrace}
\alias{matrixtrace}
\title{Trace of a matrix}
\usage{
matrixtrace(A)
}
\arguments{
\item{A}{square numeric matrix}
}
\value{
trace of a matrix
}
\description{
Trace of a matrix
}
\keyword{internal}
