% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMatrix.R
\docType{package}
\name{MixMatrix-package}
\alias{MixMatrix}
\alias{MixMatrix-package}
\title{Classification with Matrix Variate Normal and t Distributions}
\description{
Provides sampling and density functions for matrix
variate normal, \eqn{t}, and inverted \eqn{t} distributions;
ML estimation for matrix variate normal and \eqn{t} distributions
using the EM algorithm, including some restrictions on the parameters;
and classification by linear and quadratic discriminant
analysis for matrix variate normal and t distributions described
in \href{https://arxiv.org/abs/1907.09565}{Thompson et al. (2019)}.
Performs clustering with matrix variate normal and t mixture models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gzt/MixMatrix/}
  \item \url{https://gzt.github.io/MixMatrix/}
  \item Report bugs at \url{https://github.com/gzt/MixMatrix/issues}
}

}
\author{
\strong{Maintainer}: Geoffrey Thompson \email{gzthompson@gmail.com} (\href{https://orcid.org/0000-0003-2436-8822}{ORCID})

Other contributors:
\itemize{
  \item B. D. Ripley \email{ripley@stats.ox.ac.uk} (author of original lda and qda functions) [contributor, copyright holder]
  \item W. N. Venables (author of original lda and qda functions) [contributor, copyright holder]
}

}
\keyword{internal}
