% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMatrix.R
\docType{package}
\name{MixMatrix}
\alias{MixMatrix}
\alias{MixMatrix-package}
\title{Functions for Matrix Distributions}
\description{
The MixMatrix package current provides for generating
matrix variate random variables, computing matrix variate densities,
and fitting the MLEs for parameters of matrix variate normal distributions.
Currently, normal, t, and inverted t distributions are supported for generation
and density computation.
}
