% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{mle.matrixnorm}
\alias{mle.matrixnorm}
\title{mle.matrixnorm:}
\usage{
mle.matrixnorm(data, row.mean = FALSE, col.mean = FALSE,
  row.variance = "none", col.variance = "none", tol = 1e-09,
  max.iter = 100, U, V)
}
\arguments{
\item{data}{Either a list of matrices or a 3-D array with matrices in
dimensions 2 and 3, indexed by dimension 1.}

\item{row.mean}{By default, \code{FALSE}. If \code{TRUE}, will fit a
common mean within each row. If both this and \code{col.mean} are
\code{TRUE}, there will be a common mean for the entire matrix.}

\item{col.mean}{By default, \code{FALSE}. If \code{TRUE}, will fit a
common mean within each row. If both this and \code{row.mean} are
\code{TRUE}, there will be a common mean for the entire matrix.}

\item{row.variance}{Imposes a variance structure on the rows. Either
'none' or 'AR(1)'. Only positive correlations are allowed for AR(1).}

\item{col.variance}{Imposes a variance structure on the columns.
Either 'none' or 'AR(1)'. Only positive correlations are allowed for
AR(1).}

\item{tol}{Convergence criterion. Measured against square deviation
between iterations of the two variance-covariance matrices.}

\item{max.iter}{Maximum possible iterations of the algorithm.}

\item{U}{(optional) Can provide a starting point for the U matrix.
By default, an identity matrix.}

\item{V}{(optional) Can provide a starting point for the V matrix.
By default, an identity matrix.}
}
\value{
Returns a list with a mean matrix, a \eqn{U} matrix, a \eqn{V}
   matrix, the number of iterations, and error at the time of stopping.
}
\description{
Maximum likelihood estimation for matrix normal distributions
}
\examples{
set.seed(20180202)
A <- rmatrixnorm(n=100,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),
   L=matrix(c(2,1,0,.1),nrow=2),list=TRUE)
results=mle.matrixnorm(A)
print(results)


}
