% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{ones}
\alias{ones}
\title{Matrix of ones}
\usage{
ones(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}
}
\value{
Matrix of ones
}
\description{
Matrix of ones
}
\examples{
\dontrun{A <- ones(5,3)
print(A)}
}
\keyword{internal}
