% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{posmatsqrt}
\alias{posmatsqrt}
\title{Positive Matrix Square Root}
\usage{
posmatsqrt(A)
}
\arguments{
\item{A}{positive definite p x p real-valued matrix.}
}
\value{
a symmetric square root matrix for A. ie, \eqn{B = t(B)} and
   \eqn{B \%*\% B = A}.
}
\description{
Computes a positive symmetric square root  matrix for a
   positive definite input matrix. Used in the inverted matrix variate
   t-distribution.
}
\examples{
A = diag(5) + 1
B = posmatsqrt(A)
sum(abs(B - t(B)))
sum(abs(A - B \%*\% B))
}
\seealso{
\link{posmatsqrtinv}
}
