% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{posmatsqrtinv}
\alias{posmatsqrtinv}
\title{Positive Matrix Square Root Inverse}
\usage{
posmatsqrtinv(A)
}
\arguments{
\item{A}{positive definite \eqn{p \times p}{p * } real-valued matrix.}
}
\value{
a symmetric square root matrix for A. ie, \eqn{B = t(B)} and
   \eqn{B \%*\% B = A}.
}
\description{
Computes the inverse of a positive symmetric square root  matrix for a
   positive definite input matrix. Used in the inverted matrix variate
   t-distribution.
}
\examples{
A = diag(5) + 1
\dontrun{
B = posmatsqrt(A)
C = posmatsqrtinv(A)
sum(abs(B - t(B)))
sum(abs(A - B \%*\% B))
B \%*\% C}
}
\seealso{
\code{\link{posmatsqrt}}
}
\keyword{internal}
