% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matnormlda.R
\name{predict.matrixqda}
\alias{predict.matrixqda}
\title{Classify Matrix Variate Observations by Quadratic Discrimination}
\usage{
\method{predict}{matrixqda}(object, newdata, prior = object$prior, ...)
}
\arguments{
\item{object}{object of class \code{matrixqda}}

\item{newdata}{array or list of new observations to be classified.
If newdata is missing, an attempt will be made to retrieve the
data used to fit the \code{matrixqda} object.}

\item{prior}{The prior probabilities of the classes, by default the
proportions in the training set or what was set in the call to
\code{matrixqda}.}

\item{...}{arguments based from or to other methods}
}
\value{
Returns a list containing
   the following components:
   \describe{
      \item{\code{class}}{The MAP classification (a factor)}
      \item{\code{posterior}}{posterior probabilities for the classes}
   }
}
\description{
Classify matrix variate observations in conjunction with \code{matrixqda}.
}
\details{
This function is a method for the generic function \code{predict()} for
class "\code{matrixqda}". It can be invoked by calling \code{predict(x)} for
an object \code{x} of the appropriate class.
}
\examples{
set.seed(20180221)
A <- rmatrixnorm(30,mean=matrix(0,nrow=3,ncol=4))
B <- rmatrixnorm(30,mean=matrix(1,nrow=3,ncol=4))
C <- array(c(A,B), dim=c(3,4,60))
groups <- c(rep(1,30),rep(2,30))
prior <- c(.5,.5)
D <- matrixqda(C, groups, prior)
predict(D)$posterior[1:10,]

}
