% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixt.R
\name{rmatrixinvt.one}
\alias{rmatrixinvt.one}
\title{rmatrixinvt.one}
\usage{
rmatrixinvt.one(df, mean = matrix(0, nrow = 2, ncol = 2), Usqrt, Vsqrt)
}
\arguments{
\item{df}{degrees of freedom (\eqn{>0}, maybe non-integer),
\code{df = Inf} is allowed.}

\item{mean}{\eqn{p X q} This is really a 'shift' rather than a
mean as this is a central T, though the expected value will be equal
to this if \eqn{df > 2}}

\item{Usqrt}{\eqn{LL^T}  - \eqn{p X p}  positive definite symm sqrt matrix for rows,
computed from \eqn{L} if not specified. Note this is not the variance.}

\item{Vsqrt}{\eqn{R^T R}  - \eqn{q X q}  positive definite symm sqrt matrix for columns,
computed from \eqn{R}  if not specified.  Note this is not the variance.}
}
\value{
Returns a matrix of one observation. This function is for
   internal use only.
}
\description{
Generate random draws from the inverted matrix variate
   t-distribution.
}
\keyword{internal}
