% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{rmatrixnorm.one}
\alias{rmatrixnorm.one}
\title{rmatrixnorm.one}
\usage{
rmatrixnorm.one(mean, L, R)
}
\arguments{
\item{mean}{\eqn{p X q}  matrix of means}

\item{L}{\eqn{p X p}  matrix specifying relations among the rows.
By default (inherited from above), an identity matrix. Should be
output of Cholesky decomposition from rmatrixnorm.}

\item{R}{\eqn{q X q}  matrix specifying relations among the columns.
By default (inherited from above), an identity matrix. Should be
output of Cholesky decomposition from rmatrixnorm.}
}
\value{
Returns a matrix of one observation. This function is for internal use only.
}
\description{
rmatrixnorm.one
}
\keyword{internal}
