% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixt.R
\name{rmatrixt.one}
\alias{rmatrixt.one}
\title{rmatrixt.one}
\usage{
rmatrixt.one(df, mean = matrix(0, nrow = 2, ncol = 2), solveU, cholV)
}
\arguments{
\item{df}{degrees of freedom (\eqn{>0}, maybe non-integer),
\code{df = Inf} is allowed.}

\item{mean}{\eqn{p X q} This is really a 'shift' rather than a mean as this
is a central T, though the expected value will be equal to this if
\eqn{df > 2}}

\item{solveU}{\eqn{LL^T}  - \eqn{p X p}  the inverse of \eqn{U},
passed from \code{rmatrixt}.}

\item{cholV}{\eqn{R^T R = V}  - \eqn{q X q}  Cholesky decomposition of \eqn{V}}
}
\value{
Returns a matrix of one observation. This function is for internal
   use only.
}
\description{
rmatrixt.one
}
\keyword{internal}
