% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{varderiv}
\alias{varderiv}
\title{Derivative selector for chosen covariance matrix.}
\usage{
varderiv(n, rho, variance)
}
\arguments{
\item{n}{dimensions}

\item{rho}{off-diagonal parameter}

\item{variance}{variance structure - AR(1) or CS.}
}
\value{
The derivative of the selected matrix structure.
}
\description{
Derivative selector for chosen covariance matrix.
}
\keyword{internal}
