% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{varinv}
\alias{varinv}
\title{Inverse selector for chosen covariance matrix.}
\usage{
varinv(n, rho, deriv, variance)
}
\arguments{
\item{n}{dimensions}

\item{rho}{off-diagonal parameter}

\item{deriv}{logical whether to return the inverse or the derivative of
the inverse}

\item{variance}{variance structure - AR(1) or CS.}
}
\value{
The inverse or derivative of the inverse of the selected matrix structure.
}
\description{
Inverse selector for chosen covariance matrix.
}
\keyword{internal}
