% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{varmatgenerate}
\alias{varmatgenerate}
\title{varmatgenerate: selects variance structure to generate.}
\usage{
varmatgenerate(n, rho, variance)
}
\arguments{
\item{n}{number of dimensions}

\item{rho}{parameter for selected variance structure.}

\item{variance}{variance structure - AR(1) or CS.}
}
\value{
Specified matrix structure
}
\description{
varmatgenerate: selects variance structure to generate.
}
\examples{
\dontrun{varmatgenerate(5,.2,"AR(1)")}

}
\keyword{internal}
