% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{CSdet}
\alias{CSdet}
\title{Determinant for an CS covariance matrix.}
\usage{
CSdet(n, rho, deriv = FALSE)
}
\arguments{
\item{n}{dimensions}

\item{rho}{off-diagonal parameter}

\item{deriv}{logical whether to return the determinant or the
derivative of the log determinant of the inverse}
}
\value{
determinant of an CS covariance matrix
}
\description{
Determinant for an CS covariance matrix.
}
\examples{
\dontrun{
CSdet(ARgenerate(5,.4,deriv=TRUE))
}

}
\keyword{internal}
