% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{derivAR}
\alias{derivAR}
\title{Derivative of AR(1) covariance matrix}
\usage{
derivAR(n, rho)
}
\arguments{
\item{n}{dimensions}

\item{rho}{off-diagonal parameter}
}
\value{
matrix derivative of an AR(1) matrix
}
\description{
Derivative of AR(1) covariance matrix
}
\keyword{internal}
