% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{derivCS}
\alias{derivCS}
\title{Derivative of compound symmetric covariance matrix}
\usage{
derivCS(n, rho)
}
\arguments{
\item{n}{dimensions}

\item{rho}{off-diagonal parameter}
}
\description{
Derivative of compound symmetric covariance matrix
}
