% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixdist.R
\docType{package}
\name{matrixdist}
\alias{matrixdist}
\alias{matrixdist-package}
\title{matrixdist: a package for some random matrix distributions.}
\description{
The matrixdist package current provides for generating
matrix variate random variables, computing matrix variate densities,
and fitting the MLEs for parameters of matrix variate normal distributions.
Currently, normal, t, and inverted t distributions are supported for generation
and density computation.
}
