% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMatrix.R
\docType{package}
\name{MixMatrix}
\alias{MixMatrix}
\alias{MixMatrix-package}
\title{Functions for Matrix Distributions}
\description{
Provides sampling and density functions for matrix
variate normal, t, and inverted t distributions;  ML fitting for matrix
variate normal and t distributions including some restrictions on the
parameters; and linear and quadratic discriminant analysis for matrix
variate normal and t distributions. Performs mixture modeling for matrix
variate distributions.
}
