% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmatrixt.R
\name{MLmatrixt}
\alias{MLmatrixt}
\title{MLmatrixt:}
\usage{
MLmatrixt(data, row.mean = FALSE, col.mean = FALSE,
  row.variance = "none", col.variance = "none", df = 10,
  fixed = TRUE, tol = .Machine$double.eps^0.5, max.iter = 5000, U, V,
  ...)
}
\arguments{
\item{data}{Either a list of matrices or a 3-D array with matrices in
dimensions 1 and 2, indexed by dimension 3.}

\item{row.mean}{By default, \code{FALSE}. If \code{TRUE}, will fit a
common mean within each row. If both this and \code{col.mean} are
\code{TRUE}, there will be a common mean for the entire matrix.}

\item{col.mean}{By default, \code{FALSE}. If \code{TRUE}, will fit a
common mean within each row. If both this and \code{row.mean} are
\code{TRUE}, there will be a common mean for the entire matrix.}

\item{row.variance}{Imposes a variance structure on the rows. Either
'none', 'AR(1)', 'CS' for 'compound symmetry', or 'Independence' for
independent and identical variance across the rows.
Only positive correlations are allowed for AR(1) and CS.
Note that while maximum likelihood estimators are available (and used) for
the unconstrained variance matrices, \code{optim} is used for any
constraints so it may be considerably slower.}

\item{col.variance}{Imposes a variance structure on the columns.
Either 'none', 'AR(1)', 'CS', or 'Independence'. Only positive correlations are allowed for
AR(1) and CS.}

\item{df}{Starting value for the degrees of freedom. If fixed = TRUE, then this is required and not updated. By default, set to 10.}

\item{fixed}{Whether nu is estimated or fixed. By default, TRUE.}

\item{tol}{Convergence criterion. Measured against square deviation
between iterations of the two variance-covariance matrices.}

\item{max.iter}{Maximum possible iterations of the algorithm.}

\item{U}{(optional) Can provide a starting point for the U matrix.
By default, an identity matrix.}

\item{V}{(optional) Can provide a starting point for the V matrix.
By default, an identity matrix.}

\item{...}{(optional) additional arguments can be passed to \code{optim}
if using restrictions on the variance.}
}
\value{
Returns a list with a mean matrix, a \eqn{U} matrix, a \eqn{V}
   matrix, the variance parameter (the first entry of the variance matrices
   are constrained to be 1 for uniqueness), the number of iterations, the squared difference
   between iterations of the variance matrices at the time of stopping, the log likelihood,
   and a convergence code.
}
\description{
Maximum likelihood estimation for matrix normal distributions

Maximum likelihood estimates exist for \eqn{N > max(p/q,q/p)+1} and are
unique for \eqn{N > max(p,q)}. This finds the estimate for the mean and then alternates
between estimates for the \eqn{U} and \eqn{V} matrices until convergence.
An AR(1), compound symmetry, or independence restriction can be proposed for either or both
variance matrices. However, if they are inappropriate for the data, they may fail with
a warning.
}
\examples{
set.seed(20180202)
A <- rmatrixnorm(n=100,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),
   L=matrix(c(2,1,0,.1),nrow=2),list=TRUE)
results=MLmatrixnorm(A, tol = 1e-5)
print(results)


}
\seealso{
\code{\link{rmatrixnorm}}
}
