% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixt.R
\name{rmatrixinvt.one}
\alias{rmatrixinvt.one}
\title{rmatrixinvt.one}
\usage{
rmatrixinvt.one(df, mean = matrix(0, nrow = 2, ncol = 2),
  L = diag(dim(mean)[1]), R = diag(dim(mean)[2]), U = L \%*\% t(L),
  V = t(R) \%*\% R)
}
\arguments{
\item{df}{degrees of freedom (\eqn{>0}, maybe non-integer),
\code{df = Inf} is allowed.}

\item{mean}{\eqn{p X q} This is really a 'shift' rather than a
mean as this is a central T, though the expected value will be equal
to this if \eqn{df > 2}}

\item{L}{\eqn{p X p}  matrix specifying relations among the rows. By
default, an identity matrix.}

\item{R}{\eqn{q X q}  matrix specifying relations among the columns.
By default, an identity matrix.}

\item{U}{\eqn{LL^T}  - \eqn{p X p}  positive definite matrix for rows,
computed from \eqn{L} if not specified. Note this is not the variance.}

\item{V}{\eqn{R^T R}  - \eqn{q X q}  positive definite matrix for columns,
computed from \eqn{R}  if not specified.  Note this is not the variance.}
}
\value{
Returns a matrix of one observation. This function is for
   internal use only.
}
\description{
Generate random draws from the inverted matrix variate
   t-distribution.
}
\keyword{internal}
