% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{toepgenerate}
\alias{toepgenerate}
\title{toepgenerate}
\usage{
toepgenerate(n, rho)
}
\arguments{
\item{n}{number of columns/rows}

\item{rho}{correlation parameter}
}
\value{
Toeplitz $n x n$ matrix with 1 on the diagonal and $rho^k$ on
   the other diagonals, where $k$ is distance from the main diagonal.
   Used internally but it is useful for generating your own random matrices.
}
\description{
generate AR(1) correlation matrices
}
\examples{
rho <- .9
n <- 6
toepgenerate(n,rho)
}
\keyword{internal}
