% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixutils.R
\name{varmatgenerate}
\alias{varmatgenerate}
\title{Select a variance structure to generate.}
\usage{
varmatgenerate(n, rho, variance)
}
\arguments{
\item{n}{number of dimensions}

\item{rho}{parameter for selected variance structure.}

\item{variance}{variance structure - AR(1) or CS.}
}
\value{
Specified matrix structure
}
\description{
Select a variance structure to generate.
}
\keyword{internal}
